requirements.txt
blinker==1.6
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
forex-python==1.8
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
json-normalize==1.0.1
MarkupSafe==2.1.2
mongoengine==0.27.0
numpy==1.24.2
pandas==2.0.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
six==1.16.0
stripe==5.5.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.2.3
WTForms==3.0.1
XlsxWriter==3.0.9
